--- Snow Flash Tool/ConsoleApp1.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31129.286
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleApp1", "ConsoleApp1\ConsoleApp1.csproj", "{011454F3-DB1C-4D74-9D09-1CD44F0997C1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{011454F3-DB1C-4D74-9D09-1CD44F0997C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{011454F3-DB1C-4D74-9D09-1CD44F0997C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{011454F3-DB1C-4D74-9D09-1CD44F0997C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{011454F3-DB1C-4D74-9D09-1CD44F0997C1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C45CC45D-CC6E-46B9-A389-988174E8D03E}
-	EndGlobalSection
-EndGlobal
