--- Snow Flash Tool/ConsoleApp1/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Colorful.Console" version="1.2.15" targetFramework="net472" />
-  <package id="System.IO.Compression" version="4.3.0" targetFramework="net472" />
-</packages>
