+++ Snow-Flash-Tool/Snow-Flash-Tool/doAskifformat.cs
@@ -5,16 +5,16 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
-using System.Windows.Forms;
 using System.Runtime.InteropServices;
 using System.IO.Compression;
 using System.IO;
 using System.Collections;
+
 namespace ConsoleApp1
 {
-    class doAskifreboot
+    class doAskifformat
     {
-        public static async Task AskIfReboot()
+        public static async Task doAskIfformat()
         {
             string result = Console.ReadLine();
             if (result.Equals("y", StringComparison.OrdinalIgnoreCase) || result.Equals("yes", StringComparison.OrdinalIgnoreCase))
@@ -28,7 +28,7 @@ namespace ConsoleApp1
                 cmd.StartInfo.UseShellExecute = false;
                 cmd.Start();
                 Console.ForegroundColor = ConsoleColor.Blue;
-                await cmd.StandardInput.WriteLineAsync("fastboot reboot");
+                await cmd.StandardInput.WriteLineAsync("fastboot -w");
                 await Task.Delay(340);
                 Console.ResetColor();
                 cmd.StandardInput.Flush();
@@ -36,20 +36,25 @@ namespace ConsoleApp1
                 cmd.StandardOutput.ReadToEnd();
                 Console.WriteLine(cmd.StandardOutput.ReadToEnd());
                 Console.ResetColor();
-                Console.WriteLine("Finished, if you would like to return to the main menu, press any key.");
-                Console.ReadKey();
-                await Program.FlashMenu();
+                Console.ForegroundColor = ConsoleColor.DarkYellow;
+                Console.WriteLine("-------------------------------------------");
+                Console.WriteLine("| Finished. Would you like to reboot? Y/N |");
+                Console.WriteLine("-------------------------------------------");
+                Console.ResetColor();
+                await doAskifreboot.AskIfReboot();
             }
             else if (result.Equals("n", StringComparison.OrdinalIgnoreCase) || result.Equals("no", StringComparison.OrdinalIgnoreCase))
             {
                 Console.ForegroundColor = ConsoleColor.DarkRed;
-                Console.WriteLine("Skipped Reboot!");
-                await Task.Delay(340);
+                Console.WriteLine("Skipped Format!");
                 Console.ResetColor();
-                Console.WriteLine("   ");
-                Console.WriteLine("Finished, if you would like to return to the main menu, press any key.");
-                Console.ReadKey();
-                await Mainmenu.Main();
+                await Task.Delay(340);
+                Console.ForegroundColor = ConsoleColor.DarkYellow;
+                Console.WriteLine("-------------------------------------------");
+                Console.WriteLine("| Finished. Would you like to reboot? Y/N |");
+                Console.WriteLine("-------------------------------------------");
+                Console.ResetColor();
+                await doAskifreboot.AskIfReboot();
             }
             else
             {
@@ -63,10 +68,10 @@ namespace ConsoleApp1
                 Program.ClearLastLine();
                 Console.ForegroundColor = ConsoleColor.DarkYellow;
                 Console.WriteLine("---------------------------------");
-                Console.WriteLine("| Would you like to reboot? Y/N |");
+                Console.WriteLine("| Would you like to format? Y/N |");
                 Console.WriteLine("---------------------------------");
                 Console.ResetColor();
-                await AskIfReboot();
+                await doAskIfformat();
             }
         }
     }
