+++ Snow-Flash-Tool/Snow-Flash-Tool/doFormatdata.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
-using System.Windows.Forms;
 using System.Runtime.InteropServices;
 using System.IO.Compression;
 using System.IO;
@@ -13,38 +12,38 @@ using System.Collections;
 
 namespace ConsoleApp1
 {
-    class doReboottorecovery
+    class doFormatdata
     {
-        public static async Task doRebootToreocvery()
+        public static async Task doFormatData()
         {
-            Process cmd4 = new Process();
+            Process cmd5 = new Process();
 
-            cmd4.StartInfo.FileName = "cmd.exe";
-            cmd4.StartInfo.RedirectStandardInput = true;
-            cmd4.StartInfo.RedirectStandardOutput = true;
-            cmd4.StartInfo.CreateNoWindow = false;
-            cmd4.StartInfo.UseShellExecute = false;
-            Console.Title = "Selected - Reboot to recovery";
+            cmd5.StartInfo.FileName = "cmd.exe";
+            cmd5.StartInfo.RedirectStandardInput = true;
+            cmd5.StartInfo.RedirectStandardOutput = true;
+            cmd5.StartInfo.CreateNoWindow = false;
+            cmd5.StartInfo.UseShellExecute = false;
+            Console.Title = "Selected - Format Data";
             Console.Clear();
             Console.WriteLine("Are you sure you would like to continue? Y/N or R (return)");
-            string result4 = Console.ReadLine();
-            if (result4.Equals("y", StringComparison.OrdinalIgnoreCase) || result4.Equals("yes", StringComparison.OrdinalIgnoreCase))
+            string result5 = Console.ReadLine();
+            if (result5.Equals("y", StringComparison.OrdinalIgnoreCase) || result5.Equals("yes", StringComparison.OrdinalIgnoreCase))
             {
-                Console.Title = "Reboot to recovery";
+                Console.Title = "Format Data";
                 Console.Clear();
-                cmd4.Start();
+                cmd5.Start();
 
 
                 Console.ForegroundColor = ConsoleColor.DarkYellow;
-                Console.WriteLine("Rebooting");
+                Console.WriteLine("Wiping");
                 Console.ResetColor();
                 Console.WriteLine("");
                 Console.WriteLine("Make sure the device stays connected!");
                 Console.WriteLine("If any issues or unexpected problems occur, message me on Telegram @SnowTalking or xda!");
                 Console.WriteLine("");
                 Console.ForegroundColor = ConsoleColor.Blue;
-                Console.Title = "Rebooting";
-                await cmd4.StandardInput.WriteLineAsync("fastboot reboot recovery");
+                Console.Title = "Wiping";
+                await cmd5.StandardInput.WriteLineAsync("fastboot erase userdata");
                 await Task.Delay(450);
                 Console.Title = "Finished";
                 Console.ForegroundColor = ConsoleColor.DarkYellow;
@@ -55,7 +54,7 @@ namespace ConsoleApp1
                 Console.ReadKey();
                 await Mainmenu.Main();
             }
-            else if (result4.Equals("n", StringComparison.OrdinalIgnoreCase) || result4.Equals("no", StringComparison.OrdinalIgnoreCase))
+            else if (result5.Equals("n", StringComparison.OrdinalIgnoreCase) || result5.Equals("no", StringComparison.OrdinalIgnoreCase))
             {
                 Console.ForegroundColor = ConsoleColor.DarkRed;
                 Console.WriteLine("Cancelled, Closing...");
@@ -63,7 +62,7 @@ namespace ConsoleApp1
                 Environment.Exit(0);
                 Console.ReadKey();
             }
-            else if (result4.Equals("r", StringComparison.OrdinalIgnoreCase) || result4.Equals("return", StringComparison.OrdinalIgnoreCase))
+            else if (result5.Equals("r", StringComparison.OrdinalIgnoreCase) || result5.Equals("return", StringComparison.OrdinalIgnoreCase))
             {
                 await Toolsmenu.ToolsMenu();
             }
@@ -73,7 +72,7 @@ namespace ConsoleApp1
                 Console.WriteLine("Please choose a valid option!");
                 Console.ResetColor();
                 await Task.Delay(1350);
-                await doRebootToreocvery();
+                await doFormatData();
             }
         }
     }
