+++ Snow-Flash-Tool/Snow-Flash-Tool/doGetunlockdata.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
-using System.Windows.Forms;
 using System.Runtime.InteropServices;
 using System.IO.Compression;
 using System.IO;
