+++ Snow-Flash-Tool/Snow-Flash-Tool/doLockbootloader.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
-using System.Windows.Forms;
 using System.Runtime.InteropServices;
 using System.IO.Compression;
 using System.IO;
@@ -13,38 +12,39 @@ using System.Collections;
 
 namespace ConsoleApp1
 {
-    class doFormatdata
+    class doLockbootloader
     {
-        public static async Task doFormatData()
+        public static async Task doLockBootloader()
         {
-            Process cmd5 = new Process();
+            Process cmd1 = new Process();
 
-            cmd5.StartInfo.FileName = "cmd.exe";
-            cmd5.StartInfo.RedirectStandardInput = true;
-            cmd5.StartInfo.RedirectStandardOutput = true;
-            cmd5.StartInfo.CreateNoWindow = false;
-            cmd5.StartInfo.UseShellExecute = false;
-            Console.Title = "Selected - Format Data";
+            cmd1.StartInfo.FileName = "cmd.exe";
+            cmd1.StartInfo.RedirectStandardInput = true;
+            cmd1.StartInfo.RedirectStandardOutput = true;
+            cmd1.StartInfo.CreateNoWindow = false;
+            cmd1.StartInfo.UseShellExecute = false;
             Console.Clear();
-            Console.WriteLine("Are you sure you would like to continue? Y/N or R (return)");
-            string result5 = Console.ReadLine();
-            if (result5.Equals("y", StringComparison.OrdinalIgnoreCase) || result5.Equals("yes", StringComparison.OrdinalIgnoreCase))
+            Console.ForegroundColor = ConsoleColor.DarkRed;
+            Console.Title = "Selected - Lock Bootloader";
+            Console.WriteLine("WARNING! This is a dangerous command, use at your own risk! Are you sure you would like to continue? Y/N or R (return)");
+            Console.ResetColor();
+            string result2 = Console.ReadLine();
+            if (result2.Equals("y", StringComparison.OrdinalIgnoreCase) || result2.Equals("yes", StringComparison.OrdinalIgnoreCase))
             {
-                Console.Title = "Format Data";
+                Console.Title = "Locking bootloader";
                 Console.Clear();
-                cmd5.Start();
+                cmd1.Start();
 
 
                 Console.ForegroundColor = ConsoleColor.DarkYellow;
-                Console.WriteLine("Wiping");
+                Console.WriteLine("Locking bootloader");
                 Console.ResetColor();
                 Console.WriteLine("");
                 Console.WriteLine("Make sure the device stays connected!");
                 Console.WriteLine("If any issues or unexpected problems occur, message me on Telegram @SnowTalking or xda!");
                 Console.WriteLine("");
                 Console.ForegroundColor = ConsoleColor.Blue;
-                Console.Title = "Wiping";
-                await cmd5.StandardInput.WriteLineAsync("fastboot erase userdata");
+                await cmd1.StandardInput.WriteLineAsync("fastboot oem lock");
                 await Task.Delay(450);
                 Console.Title = "Finished";
                 Console.ForegroundColor = ConsoleColor.DarkYellow;
@@ -55,7 +55,7 @@ namespace ConsoleApp1
                 Console.ReadKey();
                 await Mainmenu.Main();
             }
-            else if (result5.Equals("n", StringComparison.OrdinalIgnoreCase) || result5.Equals("no", StringComparison.OrdinalIgnoreCase))
+            else if (result2.Equals("n", StringComparison.OrdinalIgnoreCase) || result2.Equals("no", StringComparison.OrdinalIgnoreCase))
             {
                 Console.ForegroundColor = ConsoleColor.DarkRed;
                 Console.WriteLine("Cancelled, Closing...");
@@ -63,9 +63,9 @@ namespace ConsoleApp1
                 Environment.Exit(0);
                 Console.ReadKey();
             }
-            else if (result5.Equals("r", StringComparison.OrdinalIgnoreCase) || result5.Equals("return", StringComparison.OrdinalIgnoreCase))
+            else if (result2.Equals("r", StringComparison.OrdinalIgnoreCase) || result2.Equals("return", StringComparison.OrdinalIgnoreCase))
             {
-                await Toolsmenu.ToolsMenu();
+                await Mainmenu.Main();
             }
             else
             {
@@ -73,7 +73,7 @@ namespace ConsoleApp1
                 Console.WriteLine("Please choose a valid option!");
                 Console.ResetColor();
                 await Task.Delay(1350);
-                await doFormatData();
+                await doLockBootloader();
             }
         }
     }
