+++ Snow-Flash-Tool/Snow-Flash-Tool/doReboottorecovery.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
-using System.Windows.Forms;
 using System.Runtime.InteropServices;
 using System.IO.Compression;
 using System.IO;
@@ -13,39 +12,38 @@ using System.Collections;
 
 namespace ConsoleApp1
 {
-    class doLockbootloader
+    class doReboottorecovery
     {
-        public static async Task doLockBootloader()
+        public static async Task doRebootToreocvery()
         {
-            Process cmd1 = new Process();
+            Process cmd4 = new Process();
 
-            cmd1.StartInfo.FileName = "cmd.exe";
-            cmd1.StartInfo.RedirectStandardInput = true;
-            cmd1.StartInfo.RedirectStandardOutput = true;
-            cmd1.StartInfo.CreateNoWindow = false;
-            cmd1.StartInfo.UseShellExecute = false;
+            cmd4.StartInfo.FileName = "cmd.exe";
+            cmd4.StartInfo.RedirectStandardInput = true;
+            cmd4.StartInfo.RedirectStandardOutput = true;
+            cmd4.StartInfo.CreateNoWindow = false;
+            cmd4.StartInfo.UseShellExecute = false;
+            Console.Title = "Selected - Reboot to recovery";
             Console.Clear();
-            Console.ForegroundColor = ConsoleColor.DarkRed;
-            Console.Title = "Selected - Lock Bootloader";
-            Console.WriteLine("WARNING! This is a dangerous command, use at your own risk! Are you sure you would like to continue? Y/N or R (return)");
-            Console.ResetColor();
-            string result2 = Console.ReadLine();
-            if (result2.Equals("y", StringComparison.OrdinalIgnoreCase) || result2.Equals("yes", StringComparison.OrdinalIgnoreCase))
+            Console.WriteLine("Are you sure you would like to continue? Y/N or R (return)");
+            string result4 = Console.ReadLine();
+            if (result4.Equals("y", StringComparison.OrdinalIgnoreCase) || result4.Equals("yes", StringComparison.OrdinalIgnoreCase))
             {
-                Console.Title = "Locking bootloader";
+                Console.Title = "Reboot to recovery";
                 Console.Clear();
-                cmd1.Start();
+                cmd4.Start();
 
 
                 Console.ForegroundColor = ConsoleColor.DarkYellow;
-                Console.WriteLine("Locking bootloader");
+                Console.WriteLine("Rebooting");
                 Console.ResetColor();
                 Console.WriteLine("");
                 Console.WriteLine("Make sure the device stays connected!");
                 Console.WriteLine("If any issues or unexpected problems occur, message me on Telegram @SnowTalking or xda!");
                 Console.WriteLine("");
                 Console.ForegroundColor = ConsoleColor.Blue;
-                await cmd1.StandardInput.WriteLineAsync("fastboot oem lock");
+                Console.Title = "Rebooting";
+                await cmd4.StandardInput.WriteLineAsync("fastboot reboot recovery");
                 await Task.Delay(450);
                 Console.Title = "Finished";
                 Console.ForegroundColor = ConsoleColor.DarkYellow;
@@ -56,7 +54,7 @@ namespace ConsoleApp1
                 Console.ReadKey();
                 await Mainmenu.Main();
             }
-            else if (result2.Equals("n", StringComparison.OrdinalIgnoreCase) || result2.Equals("no", StringComparison.OrdinalIgnoreCase))
+            else if (result4.Equals("n", StringComparison.OrdinalIgnoreCase) || result4.Equals("no", StringComparison.OrdinalIgnoreCase))
             {
                 Console.ForegroundColor = ConsoleColor.DarkRed;
                 Console.WriteLine("Cancelled, Closing...");
@@ -64,9 +62,9 @@ namespace ConsoleApp1
                 Environment.Exit(0);
                 Console.ReadKey();
             }
-            else if (result2.Equals("r", StringComparison.OrdinalIgnoreCase) || result2.Equals("return", StringComparison.OrdinalIgnoreCase))
+            else if (result4.Equals("r", StringComparison.OrdinalIgnoreCase) || result4.Equals("return", StringComparison.OrdinalIgnoreCase))
             {
-                await Mainmenu.Main();
+                await Toolsmenu.ToolsMenu();
             }
             else
             {
@@ -74,7 +72,7 @@ namespace ConsoleApp1
                 Console.WriteLine("Please choose a valid option!");
                 Console.ResetColor();
                 await Task.Delay(1350);
-                await doLockBootloader();
+                await doRebootToreocvery();
             }
         }
     }
